+++ opencontractserver/utils/etl.py
@@ -341,6 +341,8 @@ def build_document_export(
                 }
                 if annot.content_modalities:
                     annot_export["content_modalities"] = annot.content_modalities
+                if annot.long_description is not None:
+                    annot_export["long_description"] = annot.long_description
                 labelled_text.append(annot_export)
 
                 # Span annotations ({start, end}) don't have page-keyed structure